Move http_cache into cache folder.
