+++ cli/util/mod.rs
@@ -1,6 +1,6 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-// Note: Only add code in here that has no application specific logic
+// Note: Only add code in this folder that has no application specific logic
 pub mod checksum;
 pub mod diff;
 pub mod display;
